% Generated by roxygen2 (4.0.1): do not edit by hand
\name{InvLossIntegral}
\alias{InvLossIntegral}
\title{Calculate the Inverse for a Loss Integral}
\usage{
InvLossIntegral(fillRate, mn, sd, dist = "gamma", fast = T)
}
\arguments{
\item{mn}{the mean of the distribution}

\item{sd}{the standard deviation of the distribution}

\item{fillRate}{the probability value for which you wish to calculate the loss integral}

\item{dist}{the assumed dstribution of demand, 'gamma' (the default), 'norm' or 'lnorm'}

\item{fast}{TRUE for the faster interpolation method}
}
\description{
\code{InvLossIntegral} calculates the Inverse of a Loss Integral
for the chosen distribution using either interpolation from a grid (the default
method, faster but with a small loss of accuracy) or numerical integration (substantially
slower but slightly more accurate)
}
\examples{
# calculate the probability of no stock-outs associated with a fill-rate of
97\% # given sales ~ N(100,20)
InvLossIntegralExact(0.97, 100, 20, 'norm')
# and assuming a gamma distribution (the default)
InvLossIntegralExact(0.97, 100, 20, 'gamma')
# and assuming a lognormal distribution
InvLossIntegralExact(0.97, 100, 20, 'lnorm')
}
\seealso{
Other LossIntegrals: \code{\link{LossIntegral}}
}

