% Generated by roxygen2 (4.0.1): do not edit by hand
\name{InvLossIntegralFast}
\alias{InvLossIntegralFast}
\title{Calculate a FAST Inverse for a Loss Integral}
\usage{
InvLossIntegralFast(fillRate, mn, sd, dist = "gamma")
}
\arguments{
\item{mn}{the mean of the distribution}

\item{sd}{the standard deviation of the distribution}

\item{fillRate}{the probability value for which you wish to calculate the loss integral}

\item{dist}{the assumed dstribution of demand, 'gamma' (the default), 'norm' or 'lnorm'}
}
\description{
\code{InvLossIntegral} calculates a fast approximation of the Inverse Loss Integral
for the chosen distribution using interpolation from a grid
}

