% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LossIntegral}
\alias{LossIntegral}
\title{Calculate Loss Integral}
\usage{
LossIntegral(p, mn, sd, dist = "gamma", fast = T)
}
\arguments{
\item{mn}{the mean of the distribution}

\item{sd}{the standard deviation of the distribution}

\item{p}{the probability value for which you wish to calculate the loss integral}

\item{dist}{the assumed dstribution of demand, 'gamma' (the default), 'norm' or 'lnorm'}

\item{fast}{TRUE for the faster interpolation method}
}
\description{
\code{LossIntegral} calculatesthe Loss Integral
for the chosen distribution using either interpolation from a grid (the default method,
faster but with a small loss of accuracy or numerical integration (substantially
slower but slightly more accurate)
}
\examples{
# calculate the fill rate associated with a 80\% probability of
# no stock outs given sales ~N(100,40)
LossIntegral(0.80, 100, 40, 'norm')
# and assuming a gamma distribution (the default)
LossIntegral(0.80, 100, 40, 'gamma')
# and assuming a lognormal distribution
LossIntegral(0.80, 100, 40, 'lnorm')

# sensitivity analysis of safety stock against fill-rate for each distribution
xFillRate <- seq(.90, .999, length.out = 100)
mn <- 100
sd <- 80
plot(xFillRate, qnorm(InvLossIntegral(xFillRate, mn, sd, 'norm'),mn, sd)-mn ,  ty='l', col = 'red', xlab = 'Fill-Rate', ylab = "Safety Stock", ylim = c(0,500))
lines(xFillRate, qgamma2(InvLossIntegral(xFillRate, mn,sd, 'gamma'),mn, sd)-mn ,  ty='l', col = 'blue')
lines(xFillRate, qlnorm2(InvLossIntegral(xFillRate, mn,sd,'lnorm'),mn, sd)-mn ,  ty='l', col = 'green')
}
\seealso{
Other LossIntegrals: \code{\link{InvLossIntegral}}
}

