% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LossIntegralExact}
\alias{LossIntegralExact}
\title{Calculate an exact Loss Integral}
\usage{
LossIntegralExact(p, mn, sd, dist = "gamma")
}
\arguments{
\item{mn}{the mean of the normal distribution}

\item{sd}{the standard deviation of the normal distribution}

\item{p}{the probability value above which you wish to calculate the loss integral}

\item{dist}{the assumed dstribution of demand, 'gamma' (the default), 'norm' or 'lnorm'}
}
\description{
\code{LossIntegralExact} calculates an exact Loss Integral
using numerical integration. The iterative nature of numerical
integration can make it perform relatively slowly.
\code{\link{LossIntegralFast}} interpolates results from precalculated
grids which is faster but with a (very) small loss in accuracy.
}

