% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LossIntegralFast}
\alias{LossIntegralFast}
\title{Calculate a FAST Loss Integral}
\usage{
LossIntegralFast(p, mn, sd, dist = "gamma")
}
\arguments{
\item{mn}{the mean of the distribution}

\item{sd}{the standard deviation of the distribution}

\item{p}{the probability value for which you wish to calculate the loss integral}

\item{dist}{the assumed dstribution of demand, 'gamma' (the default), 'norm' or 'lnorm'}
}
\description{
\code{LossIntegral} calculates a fast approximation of the Loss Integral
for the chosen distribution using interpolation from a grid
}

