% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{LossIntegrals}
\alias{LossIntegrals}
\alias{LossIntegrals-package}
\title{Functions to handle loss integrals for use in safety stock calculations}
\description{
The LossIntegrals package is used in safety stock calculations to translate
between the probability of not stocking out in a replenishment cycle and the
expected fill-rate (the proportion of order quantities that can be met out of
inventory).
}
\details{
This package currently supports Normal, LogNormal and Gamma distributions.
}

