% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{LossIntegralTables}
\alias{LossIntegralTables}
\title{List of precalculated tables to enable rapid calculation of loss integrals and their inverses}
\description{
Precalculated tables are generated for Loss Integrals and their inverses.  We
can then use bicubic interpolation (from the akima package) to rapidly
estimate any other values on the 2D grid same  rapidly interpoat e
}

