% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qgamma2}
\alias{qgamma2}
\title{cover function for \code{\link{qgamma}} using mean and standard deviation
rather than shape and rate parameters}
\usage{
qgamma2(p, mn, sd)
}
\arguments{
\item{p,}{probability (0,1)}

\item{mn}{the mean of the gamma distribution}

\item{sd}{the standard deviation of the gamma distribution}
}
\description{
\code{qgamma2} calculates rate and shape paramaters from the provided mean
(mn) and standard deviation (sd) then calls the base R qgamma function
}
\examples{
# Use qgamma2 directly ....
qgamma2(p = 0.8, mn = 100, sd = 60)

# as an alternative to this
p <- 0.8; mn <-100 ; sd <- 60
shape <- (mn/sd)^2
rate <- mn/(sd^2)
qgamma(p, shape, rate)
}
\seealso{
Other utilities: \code{\link{qlnorm2}}
}

